ketsu flash when out of power, emmit sparkles depending on the amount of ketsu power

+++ Assets/Scripts/Game/EnemyController.cs
@@ -18,7 +18,7 @@ namespace Game
 
         [Header("Consuming")]
         public bool CanBeConsumed;
-        public float GivesKetsuPower;
+        public int GivesKetsuPower;
 
         [Header("Dazing")]
         public bool CanBeDazed;
